# backend/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Luma Local Media API
  version: 0.1.0
servers:
  - url: /
paths:
...
  /api/v1/system/info:
    get:
      summary: 查询服务端运行信息
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 服务端运行信息
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemInfo" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/sources:
    get:
      summary: 查询媒体源列表
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: 不包含真实根路径的媒体源列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Source" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: 创建本地媒体源
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSource" }
      responses:
        "201":
          description: 媒体源已创建
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "403": { $ref: "#/components/responses/ForbiddenPath" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/sources/{id}:
    parameters:
      - $ref: "#/components/parameters/SourceID"
    patch:
      summary: 更新或重新启用媒体源
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSource" }
      responses:
        "200":
          description: 更新后的媒体源
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: 软禁用媒体源并保留全部索引
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: 媒体源已禁用 }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/sources/{id}/scan:
    parameters:
      - $ref: "#/components/parameters/SourceID"
    post:
      summary: 创建异步媒体源扫描任务
      security: [{ bearerAuth: [] }]
      responses:
        "202":
          description: 扫描任务已持久化并等待执行
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanJob" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/scan-jobs/latest:
    get:
      summary: 查询最近扫描任务
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: source_id
          schema: { type: string }
      responses:
        "200":
          description: 最近扫描任务
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanJob" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/scan-jobs/{id}:
    get:
      summary: 查询扫描任务状态和进度
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 扫描任务详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanJob" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/media:
    get:
      summary: 查询媒体列表
      description: |
        默认排除状态为 missing 的媒体，以及已禁用或软删除来源下的媒体。
        sort=duration 时仅包含 duration 已稳定的行（duration_ms 非空，或 status 为 ready/failed）。
        sort=file_size 时仅包含 status 为 ready/failed 的行。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: q
          description: 按文件名搜索，最多 200 个字符。
          schema: { type: string, maxLength: 200 }
        - in: query
          name: type
          schema: { type: string, enum: [video, image] }
        - in: query
          name: library_kind
          description: 严格限制视频来源用途；个人视频页使用 personal，避免混入电影和电视剧源。图片库仅按 type=image 查询。
          schema: { type: string, enum: [personal, movies, tv] }
        - in: query
          name: favorite
          description: true 仅返回收藏媒体；false 包含未创建用户数据的媒体。
          schema: { type: boolean }
        - in: query
          name: tag_id
          description: 仅返回关联当前用户指定标签的媒体。
          schema: { type: string, maxLength: 200 }
        - in: query
          name: watch_status
          description: 按当前用户观看状态筛选；无用户数据的媒体归入 unwatched。
          schema: { type: string, enum: [unwatched, watching, completed] }
        - in: query
          name: sort
          description: created_at/filename 含处理中媒体；duration/file_size 仅含排序键已稳定的行。
          schema:
            type: string
            enum: [created_at, filename, duration, file_size]
            default: created_at
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc], default: desc }
        - in: query
          name: cursor
          description: 上一页返回的不透明游标；筛选或排序条件改变后不可复用。
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: 媒体列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaPage" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/media/count:
    get:
      summary: 统计当前可见媒体数量
      description: 使用与媒体列表相同的来源授权和筛选规则，仅返回聚合数量，不分页也不返回媒体条目。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: q
          schema: { type: string, maxLength: 200 }
        - in: query
          name: type
          schema: { type: string, enum: [video, image] }
        - in: query
          name: library_kind
          schema: { type: string, enum: [personal, movies, tv] }
        - in: query
          name: favorite
          schema: { type: boolean }
        - in: query
          name: tag_id
          schema: { type: string, maxLength: 200 }
        - in: query
          name: watch_status
          schema: { type: string, enum: [unwatched, watching, completed] }
      responses:
        "200":
          description: 当前用户可见媒体总数
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaCount" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/media/continue-watching:
    get:
      summary: 查询继续观看列表
      description: 返回有进度、未完成且当前可见的视频，按最近播放时间和媒体 ID 降序稳定分页。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: cursor
          description: 与当前用户和继续观看集合绑定的不透明游标。
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: 继续观看媒体列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaPage" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/media/{id}:
    parameters:
      - $ref: "#/components/parameters/MediaID"
    get:
      summary: 查询媒体详情
      description: missing 媒体及已禁用或软删除来源下的媒体视为不存在。
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: 媒体详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaDetail" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/MediaNotFound" }
  /api/v1/media/{id}/thumbnail:
    parameters:
      - $ref: "#/components/parameters/MediaID"
    get:
      summary: 获取媒体缩略图
      description: |
        返回当前可用的默认或卡片图片资产；支持使用 If-None-Match 进行 ETag 条件请求。
        若资产已记录内容哈希且 If-None-Match 命中，服务端可直接返回 304 而不读取文件体。
        内容变更重新生成完成前，仍可能提供上一份 ready 缩略图。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: variant
          description: 省略时保持原有默认缩略图行为；card 为 640×400 居中裁剪卡片图。
          schema: { type: string, enum: [default, card], default: default }
        - in: header
          name: If-None-Match
          description: 客户端缓存的缩略图 ETag。
          schema: { type: string }
      responses:
        "200":
          description: 缩略图二进制内容
          headers:
            ETag:
              description: 当前缩略图资产的强 ETag。
              schema: { type: string }
            Cache-Control:
              description: 固定为 private, max-age=604800, must-revalidate。
              schema: { type: string, const: "private, max-age=604800, must-revalidate" }
          content:
            image/*:
              schema: { type: string, format: binary }
        "304":
          description: If-None-Match 与当前 ETag 匹配，无响应体。
          headers:
            ETag:
              schema: { type: string }
            Cache-Control:
              schema: { type: string, const: "private, max-age=604800, must-revalidate" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/MediaOrThumbnailNotFound" }
  /api/v1/media/{id}/stream:
    parameters:
      - $ref: "#/components/parameters/MediaID"
    get:
      summary: 播放原始视频
      description: |
        仅服务可见的视频原文件，支持标准单段 HTTP Range、条件请求和进度跳转。
        discovered、probing、thumbnailing、ready、failed 状态均可 Direct Play；图片和不可见媒体返回 404。
        弱 ETag 与 Content-Length 均基于打开瞬间的 size/mtime 快照；媒体源根目录不可访问时返回 503 SOURCE_OFFLINE。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: Range
          schema: { type: string, examples: ["bytes=0-1048575"] }
        - in: header
          name: If-None-Match
          schema: { type: string }
        - in: header
          name: If-Modified-Since
          schema: { type: string }
        - in: header
          name: If-Range
          description: 弱 ETag 不能用于 If-Range 强比较；客户端可使用 Last-Modified 日期。
          schema: { type: string }
      responses:
        "200":
          description: 完整视频内容
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer, format: int64 } }
            ETag: { schema: { type: string }, description: 由打开瞬间文件大小和秒级修改时间生成的弱 ETag。 }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
          content:
            video/*:
              schema: { type: string, format: binary }
        "206":
          description: Range 指定的部分视频内容
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer, format: int64 } }
            Content-Range: { schema: { type: string } }
            ETag: { schema: { type: string } }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
          content:
            video/*:
              schema: { type: string, format: binary }
        "304":
          description: 条件请求命中，无响应体
          headers:
            ETag: { schema: { type: string } }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/MediaNotFound" }
        "416":
          description: Range 不可满足；Content-Range 为 bytes */快照文件大小，响应体由 Go 标准库生成。
          headers:
            Content-Range: { schema: { type: string } }
        "503":
          description: 媒体源根目录当前不可访问，error.code 为 SOURCE_OFFLINE
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    head:
      summary: 查询原始视频响应元数据
      description: 与 GET 使用相同的认证、可见性和条件请求语义，但不返回响应体。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: If-None-Match
          schema: { type: string }
        - in: header
          name: If-Modified-Since
          schema: { type: string }
      responses:
        "200":
          description: 视频响应元数据
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer, format: int64 } }
            ETag: { schema: { type: string } }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
        "304": { description: 条件请求命中，无响应体 }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/MediaNotFound" }
        "503":
          description: 媒体源根目录当前不可访问，error.code 为 SOURCE_OFFLINE
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/media/{id}/original:
    parameters:
      - $ref: "#/components/parameters/MediaID"
    get:
      summary: 浏览原始图片
      description: |
        仅服务可见的图片原文件，支持 HTTP Range 和条件请求；视频和不可见媒体返回 404。
        discovered、probing、thumbnailing、ready、failed 状态均可读取原图，不依赖缩略图处理结果。
        弱 ETag 与 Content-Length 均基于打开瞬间的 size/mtime 快照；媒体源根目录不可访问时返回 503 SOURCE_OFFLINE。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: Range
          schema: { type: string, examples: ["bytes=0-1048575"] }
        - in: header
          name: If-None-Match
          schema: { type: string }
        - in: header
          name: If-Modified-Since
          schema: { type: string }
        - in: header
          name: If-Range
          description: 弱 ETag 不能用于 If-Range 强比较；客户端可使用 Last-Modified 日期。
          schema: { type: string }
      responses:
        "200":
          description: 完整原始图片
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer, format: int64 } }
            ETag: { schema: { type: string }, description: 由打开瞬间文件大小和秒级修改时间生成的弱 ETag。 }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
            X-Content-Type-Options: { schema: { type: string, const: nosniff } }
          content:
            image/*:
              schema: { type: string, format: binary }
        "206":
          description: Range 指定的部分图片内容
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer, format: int64 } }
            Content-Range: { schema: { type: string } }
            ETag: { schema: { type: string } }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
            X-Content-Type-Options: { schema: { type: string, const: nosniff } }
          content:
            image/*:
              schema: { type: string, format: binary }
        "304":
          description: 条件请求命中，无响应体
          headers:
            ETag: { schema: { type: string } }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
            X-Content-Type-Options: { schema: { type: string, const: nosniff } }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/MediaNotFound" }
        "416":
          description: Range 不可满足；Content-Range 为 bytes */快照文件大小，响应体由 Go 标准库生成。
          headers:
            Content-Range: { schema: { type: string } }
        "503":
          description: 媒体源根目录当前不可访问，error.code 为 SOURCE_OFFLINE
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    head:
      summary: 查询原始图片响应元数据
      description: 与 GET 使用相同的认证、可见性和条件请求语义，但不返回响应体。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: If-None-Match
          schema: { type: string }
        - in: header
          name: If-Modified-Since
          schema: { type: string }
      responses:
        "200":
          description: 原始图片响应元数据
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer, format: int64 } }
            ETag: { schema: { type: string } }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
            X-Content-Type-Options: { schema: { type: string, const: nosniff } }
        "304": { description: 条件请求命中，无响应体 }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/MediaNotFound" }
        "503":
          description: 媒体源根目录当前不可访问，error.code 为 SOURCE_OFFLINE
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/media/{id}/user-data:
    parameters:
      - $ref: "#/components/parameters/MediaID"
    get:
      summary: 查询媒体用户数据
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: 当前用户的完整媒体数据；无记录时 revision 为 0。
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaUserData" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/MediaNotFound" }
    patch:
      summary: 原子修改用户数据和媒体标签
      description: 字段缺失表示保持不变；custom_title/notes 为 null 表示清除；tag_ids 表示完整替换。base_revision 不匹配返回 409。
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateMediaUserData" }
      responses:
        "200":
          description: 更新后的完整用户数据
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaUserData" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: 媒体不存在时 error.code 为 MEDIA_NOT_FOUND；tag_ids 含不存在或不属于当前用户的标签时为 TAG_NOT_FOUND。
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/media/{id}/progress:
    parameters:
      - $ref: "#/components/parameters/MediaID"
    put:
      summary: 保存视频播放进度
      description: |
        在同一事务内读取服务端媒体时长后截断 position_ms；达到时长 90% 自动完成。
        与 PATCH /user-data 共用 media_user_data.revision；旧 base_revision 返回 409 REVISION_CONFLICT。
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProgress" }
      responses:
        "200":
          description: 更新后的完整用户数据
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaUserData" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/MediaNotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422":
          description: 目标媒体不是可播放视频，error.code 为 MEDIA_NOT_PLAYABLE。
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/tags:
    get:
      summary: 查询当前用户标签
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: 标签列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Tag" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: 创建标签
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTag" }
      responses:
        "201":
          description: 已创建标签
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tag" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/tags/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    patch:
      summary: 重命名标签
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTag" }
      responses:
        "200":
          description: 更新后的标签
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tag" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      summary: 删除标签和关联关系
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: 标签已删除 }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/catalog:
    get:
      summary: 查询电影和电视剧作品库
      description: 查询前幂等整理新增或发生变化的影视文件；个人媒体不进入作品库。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: kind
          schema: { type: string, enum: [movie, series] }
        - in: query
          name: q
          schema: { type: string, maxLength: 200 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: 作品级媒体列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/CatalogItem" } }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/admin/users:
    get:
      summary: 查询用户（仅管理员）
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: 用户列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/AccessUser" } }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      summary: 创建成员（仅管理员）
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 80 }
                request_id:
                  type: string
                  maxLength: 128
                  description: 客户端为一次创建操作生成并在网络重试时复用的不透明幂等键。
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccessUser" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/users/{id}:
    patch:
      summary: 修改成员名称或启用状态（仅管理员）
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: { type: string, minLength: 1, maxLength: 80 }
                enabled: { type: boolean }
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccessUser" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/users/{id}/tokens:
    get:
      summary: 查询成员令牌元数据（仅管理员）
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: 不包含令牌明文
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/AccessToken" } }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      summary: 签发成员令牌（仅管理员，明文只返回一次）
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 80 }
                expires_at: { type: [string, "null"], format: date-time }
                request_id:
                  type: string
                  maxLength: 128
                  description: 客户端为一次签发操作生成并在网络重试时复用的不透明幂等键。
      responses:
        "201":
          description: 已签发；token 字段不会再次返回
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IssuedAccessToken" }
        "409":
          description: 令牌已创建但服务重启后无法安全重放一次性明文；不得以新 request_id 自动重试。
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/tokens/{id}:
    delete:
      summary: 吊销成员令牌（仅管理员）
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204": { description: 已吊销 }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/users/{id}/sources:
    get:
      summary: 查询成员可访问的媒体源（仅管理员）
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: 来源 ID 列表
          content:
            application/json:
              schema:
                type: object
                required: [source_ids]
                properties:
                  source_ids: { type: array, items: { type: string } }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/users/{id}/sources/{sourceId}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
      - { in: path, name: sourceId, required: true, schema: { type: string } }
    put:
      summary: 授予媒体源访问权（仅管理员）
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: 已授权 }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      summary: 移除媒体源访问权（仅管理员）
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: 已移除 }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/catalog/issues:
    get:
      summary: 查询待人工整理的影视文件
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: 未能可靠识别的文件
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/CatalogIssue" } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    SystemInfo:
      type: object
      required: [version, platform, architecture, database, user, capabilities]
      properties:
        version: { type: string }
        platform: { type: string }
        architecture: { type: string }
        database: { type: string }
        user:
          type: object
          required: [id, name, role]
          properties:
            id: { type: string }
            name: { type: string }
            role: { type: string, enum: [admin, member] }
        capabilities:
          type: array
          items: { type: string }
...
    CreateSource:
      type: object
      additionalProperties: false
      required: [name, root_path]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        root_path: { type: string, minLength: 1 }
        library_kind: { type: string, enum: [personal, movies, tv], default: personal }
...
    Source:
      type: object
      required: [id, name, type, library_kind, enabled, status, created_at, updated_at]
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string, const: local }
        library_kind: { type: string, enum: [personal, movies, tv] }
        enabled: { type: boolean }
        status: { type: string, enum: [online, offline, degraded, disabled] }
        last_scan_id: { type: string }
        last_seen_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: API Token 缺失或无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...